requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
PyPDF2
faiss-cpu